added google login section

+++ app/Http/Controllers/ItemController.php
@@ -3,11 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\ITEM;
+use Session;
+use App\User;
 
 class ItemController extends Controller
 {
     public function store(Request $request){
+        //check for correct user
+        $userId = Session::get('id');
+        $requiredUserRow = User::find($userId);
+        $role = $requiredUserRow->role;
+        if($role != 'admin'){
+            abort(404,'page not found');
+        }
         // validate the inputs
         $this->validate($request,[
             'item'=>'required|max:200|min:5',
@@ -43,7 +53,7 @@ class ItemController extends Controller
          $data = array('item'=>$varitem,'catagory'=>$varcatagory,'quality'=>$varquality,'description'=>$vardescription,'created_at'=>$createTime,'cover_image'=>$fileNameToStore);
         
         \DB::table('items')->insert($data);
-        return view('addItems')->withMessage("New Item added Successfully !!!");
+        return view('pages.home')->withMessage("New Item added Successfully !!!");
         
     }
 
@@ -53,18 +63,43 @@ class ItemController extends Controller
     }
 
     public function deteteItem($id){
+        //check for correct user
+        $userId = Session::get('id');
+        $requiredUserRow = User::find($userId);
+        $role = $requiredUserRow->role;
+        if($role != 'admin'){
+            abort(404,'page not found');
+        }
         $requiredRow = ITEM::find($id);
+        
+        if($requiredRow->cover_image != 'noimage.jpg'){
+            //delete cover_image from storage
+            Storage::delete('public/cover_images/'.$requiredRow->cover_image);
+        }
         $requiredRow->delete();
-        return redirect()->back(); 
+        return view('pages.home')->withMessage('Deleted Successfully'); 
     }
 
     public function updateItem($id){
+        //check for correct user
+        $userId = Session::get('id');
+        $requiredUserRow = User::find($userId);
+        $role = $requiredUserRow->role;
+        if($role != 'admin'){
+            abort(404,'page not found');
+        }
         $requiredRow = ITEM::find($id);
         return view('updatePage')->with('itemdata',$requiredRow);
     }
 
     public function update(Request $request){
-        
+        //check for correct user
+        $userId = Session::get('id');
+        $requiredUserRow = User::find($userId);
+        $role = $requiredUserRow->role;
+        if($role != 'admin'){
+            abort(404,'page not found');
+        }
         // validate the inputs
         $this->validate($request,[
             'item'=>'required|max:200|min:5',
@@ -88,9 +123,7 @@ class ItemController extends Controller
             //upload the image
             $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
         }
-        else{
-            $fileNameToStore = 'noimage.jpg';
-        }
+        
     
         $updateTime = new \DateTime();
 
@@ -101,10 +134,12 @@ class ItemController extends Controller
         $requiredRow->quality = $request->input('quality');
         $requiredRow->description = $request->input('description');
         $requiredRow->updated_at = $request->input('created_at');
-        $requiredRow-> cover_image = $fileNameToStore;
+        if($request->hasFile('cover_image')){
+            $requiredRow->cover_image = $fileNameToStore;
+        }
         $requiredRow->save();
         
-        return redirect()->back()->withMessage('Updated Successfully');
+        return view('pages.home')->withMessage('Updated Successfully');
         
     }
 }
